+++ fabric-sponge-command-api-v2/src/main/java/net/fabricmc/fabric/api/command/v2/lib/sponge/args/ChildCommandElementExecutor.java
@@ -41,7 +41,7 @@ import net.minecraft.text.LiteralText;
 import net.minecraft.text.Text;
 import net.minecraft.world.World;
 
-import net.fabricmc.fabric.api.command.v2.Location;
+import net.fabricmc.fabric.api.util.Location;
 import net.fabricmc.fabric.api.command.v2.PermissibleCommandSource;
 import net.fabricmc.fabric.api.command.v2.lib.sponge.CommandCallable;
 import net.fabricmc.fabric.api.command.v2.lib.sponge.CommandException;
